Changed Location#getLongitude and Location#getLatitude to return double's rather than long's to be safe.

+++ api/src/main/java/org/telegram/botapi/api/chat/message/content/type/Location.java
@@ -5,7 +5,7 @@ package org.telegram.botapi.api.chat.message.content.type;
  */
 public interface Location {
 
-    float getLongitude();
+    double getLongitude();
 
-    float getLatitude();
+    double getLatitude();
 }
